Try sending image from Rust to Swift
Doesn't compile, but I think this is in the right direction.

+++ libloadviz/src/lib.rs
@@ -7,8 +7,8 @@ pub extern "C" fn add(left: i32, right: i32) -> i32 {
 }
 
 #[no_mangle]
-pub extern "C" fn get_image(width: u16, height: u16) -> *const u8 {
-    let image: Vec<u8> = vec![0; (width as usize) * (height as usize) * 3];
+pub extern "C" fn get_image(width: usize, height: usize) -> *const u8 {
+    let image: Vec<u8> = vec![0; width * height * 3];
     let boxed_image = image.into_boxed_slice();
 
     // FIXME: Make sure this memory isn't just leaked!
